requirements.txt
Django>=4.2
gunicorn==22.0.0
psycopg2-binary==2.9.9
dj_database_url==2.2.0
python-dotenv==1.0.1
matplotlib==3.9.2
openai==1.50.1
google-generativeai==0.8.3
reportlab==4.2.2
pypdf==5.0.1
python-docx==1.1.2
xhtml2pdf>=0.2.11
pdfminer.six>=20221105
beautifulsoup4>=4.12.3
PyMuPDF>=1.24.7
docx2txt>=0.8
